+++ packages/db/src/search/BaseSearch.ts
@@ -185,9 +185,15 @@ class BaseSearch {
         case 'pdrName':
           [countQuery, searchQuery].forEach((query) => query?.[queryMethod](`${this.tableName}.pdr_cumulus_id`));
           break;
+        case 'asyncOperationId':
+          [countQuery, searchQuery].forEach((query) => query?.[queryMethod](`${this.tableName}.async_operation_cumulus_id`));
+          break;
         case 'error':
         case 'error.Error':
           [countQuery, searchQuery].forEach((query) => query?.whereRaw(`${this.tableName}.error ->> 'Error' is ${checkNull}`));
+          break;
+        case 'parentArn':
+          [countQuery, searchQuery].forEach((query) => query?.[queryMethod](`${this.tableName}.parent_cumulus_id`));
           break;
         default:
           [countQuery, searchQuery].forEach((query) => query?.[queryMethod](`${this.tableName}.${name}`));
@@ -241,6 +247,8 @@ class BaseSearch {
       collections: collectionsTable,
       providers: providersTable,
       pdrs: pdrsTable,
+      asyncOperations: asyncOperationsTable,
+      executions: executionsTable,
     } = TableNames;
 
     const { countQuery, searchQuery, dbQueryParameters } = params;
@@ -263,6 +271,12 @@ class BaseSearch {
         case 'error.Error':
           [countQuery, searchQuery]
             .forEach((query) => query?.whereRaw(`${this.tableName}.error->>'Error' = '${value}'`));
+          break;
+        case 'asyncOperationId':
+          [countQuery, searchQuery].forEach((query) => query?.where(`${asyncOperationsTable}.id`, value));
+          break;
+        case 'parentArn':
+          [countQuery, searchQuery].forEach((query) => query?.where(`${executionsTable}.arn`, value));
           break;
         default:
           [countQuery, searchQuery].forEach((query) => query?.where(`${this.tableName}.${name}`, value));
@@ -288,6 +302,8 @@ class BaseSearch {
       collections: collectionsTable,
       providers: providersTable,
       pdrs: pdrsTable,
+      asyncOperations: asyncOperationsTable,
+      executions: executionsTable,
     } = TableNames;
 
     const { countQuery, searchQuery, dbQueryParameters } = params;
@@ -319,6 +335,12 @@ class BaseSearch {
           [countQuery, searchQuery]
             .forEach((query) => query?.whereRaw(`${this.tableName}.error->>'Error' in ('${value.join('\',\'')}')`));
           break;
+        case 'asyncOperationId':
+          [countQuery, searchQuery].forEach((query) => query?.whereIn(`${asyncOperationsTable}.id`, value));
+          break;
+        case 'parentArn':
+          [countQuery, searchQuery].forEach((query) => query?.whereIn(`${executionsTable}.arn`, value));
+          break;
         default:
           [countQuery, searchQuery].forEach((query) => query?.whereIn(`${this.tableName}.${name}`, value));
           break;
@@ -343,6 +365,7 @@ class BaseSearch {
       collections: collectionsTable,
       providers: providersTable,
       pdrs: pdrsTable,
+      asyncOperations: asyncOperationsTable,
     } = TableNames;
 
     const { countQuery, searchQuery, dbQueryParameters } = params;
@@ -362,6 +385,9 @@ class BaseSearch {
           break;
         case 'pdrName':
           [countQuery, searchQuery].forEach((query) => query?.whereNot(`${pdrsTable}.name`, value));
+          break;
+        case 'asyncOperationId':
+          [countQuery, searchQuery].forEach((query) => query?.whereNot(`${asyncOperationsTable}.id`, value));
           break;
         case 'error.Error':
           [countQuery, searchQuery].forEach((query) => query?.whereRaw(`${this.tableName}.error->>'Error' != '${value}'`));