+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 ### Replace ElasticSearch Phase 1
+- **CUMULUS-3239**
+ - Updated `execution` list api endpoint and added `ExecutionSearch` class to query postgres
 - **CUMULUS-3699**
  - Updated `collections` api endpoint to be able to support `includeStats` query string parameter
 - **CUMULUS-3641**